style: fix ruff lint + apply ruff format
CI has been failing the Lint step on every Phase 1 commit. Resolved:

- Per-file B008 ignore for src/selexprep/cli.py (Typer's documented
  idiom calls typer.Option/typer.Argument as defaults; B008 is a
  false positive for Typer-style CLI definitions)
- demux.py: replaced E702 semicolon-chained writes with one-per-line
  and added SIM115 noqa for the writer-factory pattern (handles are
  closed in the outer `finally`, not a context-manager case)
- discover.py: B023 closure binding fixed via default-arg capture
  (`_cols=cols`); RUF012 ClassVar annotations on adapter QUERIES lists;
  B007 unused loop var renamed to `_bp_id`; SIM103 inlined; SIM105
  replaced try/except/pass with contextlib.suppress
- audit.py: RUF002 en-dash → hyphen in docstring
- detect.py: RUF046 redundant int(round(...)) → round(...)
- test_readiness.py: RUF059 `_std` for unused unpacked tuple slot
- Applied `ruff format` across all 32 src + test files (20 reformatted)

166 tests still pass. ruff check + ruff format --check now both clean,
which is exactly what .github/workflows/tests.yml requires.

+++ src/selexprep/library/detect.py
@@ -167,7 +167,7 @@ def detect_primers(
     p5 = detect_flank(sequences, is_prefix=True, **common_kwargs)
     p3 = detect_flank(sequences, is_prefix=False, **common_kwargs)
 
-    mean_len = int(round(pd.Series([len(s) for s in sequences]).mean()))
+    mean_len = round(pd.Series([len(s) for s in sequences]).mean())
     random_len: int | None = None
     if p5.sequence or p3.sequence:
         random_len = mean_len - p5.length - p3.length
